Added basic achievements menu with persistent achievements (#2)
Initial prototype for an achievements menu until we have a more concrete
design.

Achievement menu can be accessed from the start menu (before game is
loaded) or in the pause menu. Currently just includes achievement name
and description as well as number of completed achievements. Example
usage can be seen in the script.

Achievements progress will be saved in a file in the save directory;
this means that they should persist across multiple saves and between
the game opening and closing, so different achievements can be
accomplished on different savefiles.

Screenshot:

![achievements_menu](https://github.com/athamante/noctyx_before_corruption/assets/152583383/55643e54-2ea0-4164-88da-8c8b4afac72e)