Added dress-up-game prototype with placeholder image (#3)
Adds a new prototype screen for the dress-up minigame as well as the
code for combining components into a single sprite for Uki. Note that
since we don't actually have any assets yet, that code is untested and
may not be functional.

Couldn't test the actual combining, but did check the variables and
clicking Prev/Next on the dress-up screen properly updates the state for
what outfit components Uki should be using.

This makes some assumptions on filenames, such as different hair
components being named `hair0`, `hair1`, etc. in the `sprites/uki`
folder and similarly for other component types. Also has some
placeholders for component size and positioning and assumes four
variants of each right now, to be updated with proper numbers later.

Screenshot (this is placeholder Uki image):

![dress_up_prototype](https://github.com/athamante/noctyx_before_corruption/assets/152583383/f80af97b-9908-4297-8b28-6f25fcabfed6)

---------

Co-authored-by: Dalia Grodzka <39556362+NiyeSadece@users.noreply.github.com>