Added name selection (#5)
Adds a naming function.

Player is asked to choose a name and if it's on one of the 'forbidden
names' list, a corresponding message is displayed and the player is
prompted to choose a name again. In special cases, the game closes.
Chosen name is correctly displayed above dialogue window.

This makes assumption on filenames for the CGs that are supposed to pops
up in some cases.

Marked as prototype because we might want to include regex or make
separate view for the naming.