Add prototype for Fulgur's information journal (#6)
Prototype for Fulgur's information screen, which currently looks pretty
much exactly like the history screen. Will include location where
information was logged + the information being logged.

When in Fulgur's POV, we'll be able to access this screen both from the
pause menu and also from a button being displayed in the top left. For
now, I'm using the Nijisanji logo (and a lightened version) as a
placeholder.

Also adds `pov.rpy` in the main folder for any general POV-switching
logic we might need, like a Takaradachi counter for Alban or a SAN
meter.

Button in top left:

![info_button_screenshot](https://github.com/athamante/noctyx_before_corruption/assets/152583383/fde0c6a1-39af-4f41-a1dd-3e95fe8b6e6f)

Information screen:

![info_screen_screenshot](https://github.com/athamante/noctyx_before_corruption/assets/152583383/99dae500-37ae-49b3-973d-042db118919c)