Add prototype for character bio screen (#7)
Protoype for character bio screen, accessible from in-game menu.

Has a row of icons at the bottom that can be selected and switched
between (currently using same icons for Noctyx member and mascot +
default placeholder images for sprites, all of which will be replaced
later). Character bio includes sprite, name, description, and for
non-mascots a mock-up of their RPG stats; this RPG portion can be
included or not when we figure out design later.

This is mostly just to have something roughly working, I imagine future
design could be quite different.

Also makes a minor fix to achievements saving.



https://github.com/athamante/noctyx_before_corruption/assets/152583383/ea6a64ec-1cd3-4136-afd0-910ce473f36d