Time based choices (#8)
Created and implemented a screen for time-limited choices with a custom
time limit and label to jump to in case the timer runs out. I had the
help of Jenn to solve a local variable problem.
Rollback is already disabled for the screen, and the display will be
modifiable later.